+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+- [debug] added support for managing debug sessions for extensions from debug panel (previously only possible using `Hosted Plugin` commands) [#8706](https://github.com/eclipse-theia/theia/pull/8706)
+
 ## v1.15.0 - 6/30/2021
 
 [1.15.0 Milestone](https://github.com/eclipse-theia/theia/milestone/21)
@@ -109,7 +111,6 @@
 - [core] updated `nsfw` dependency to `^2.1.2` [#9267](https://github.com/eclipse-theia/theia/pull/9267)
 - [debug] fixed hover issues for the `currentFrame` editor [#9256](https://github.com/eclipse-theia/theia/pull/9256)
 - [debug] improved error messages [#9386](https://github.com/eclipse-theia/theia/pull/9386)
-- [debug] added support for managing debug sessions for extensions from debug panel (previously only possible using `Hosted Plugin` commands) [#8706](https://github.com/eclipse-theia/theia/pull/8706)
 - [documentation] added roadmap information to the readme [#9308](https://github.com/eclipse-theia/theia/pull/9308)
 - [documentation] updated pre-publishing steps [#9257](https://github.com/eclipse-theia/theia/pull/9257)
 - [editor-preview] updated logic to activate editor-preview editors only if already active [#9346](https://github.com/eclipse-theia/theia/pull/9346)